updating structure for RDB
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C53651

+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhS_BE/Tables/SalePriceList_New.sql
@@ -4,7 +4,7 @@
     [OwnerID]                INT            NOT NULL,
     [CurrencyID]             INT            NOT NULL,
     [EffectiveOn]            DATETIME       NULL,
-    [PriceListType]          TINYINT        NULL,
+    [PriceListType]          INT            NULL,
     [timestamp]              ROWVERSION     NULL,
     [SourceID]               VARCHAR (50)   NULL,
     [ProcessInstanceID]      BIGINT         NULL,
@@ -15,6 +15,8 @@
     [Description]            NVARCHAR (MAX) NULL,
     [PricelistStateBackupID] BIGINT         NULL
 );
+
+
