+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhS_BE/Tables/SalePricelistCodeChange_New.sql
@@ -3,12 +3,14 @@
     [RecentZoneName] NVARCHAR (150) NULL,
     [ZoneName]       NVARCHAR (150) NULL,
     [ZoneID]         BIGINT         NULL,
-    [Change]         TINYINT        NULL,
+    [Change]         INT            NULL,
     [BatchID]        INT            NOT NULL,
     [BED]            DATETIME       NOT NULL,
     [EED]            DATETIME       NULL,
     [CountryID]      INT            NULL
 );
+
+
