+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhs_CP/Tables/SaleZoneRoutingProduct_Preview.sql
@@ -1,13 +1,15 @@
 ﻿CREATE TABLE [TOneWhs_CP].[SaleZoneRoutingProduct_Preview] (
     [ProcessInstanceID] BIGINT         NOT NULL,
     [ZoneName]          NVARCHAR (255) NOT NULL,
-    [OwnerType]         TINYINT        NOT NULL,
+    [OwnerType]         INT            NOT NULL,
     [OwnerID]           INT            NOT NULL,
     [RoutingProductID]  INT            NOT NULL,
     [BED]               DATETIME       NOT NULL,
     [EED]               DATETIME       NULL,
     [ChangeType]        INT            NOT NULL
 );
+
+
